Changing the inteface of ExpressionEvaluator. This gives a way to override expression parsing and expression evaluation logic.
